+++ benchmarks/App.Metrics.Benchmarks/BenchmarkDotNetBenchmarks/Concurrency/AtomicLongArrayBenchMark.cs
@@ -1,19 +1,19 @@
-// Copyright (c) Allan Hardy. All rights reserved.
+﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using App.Metrics.Concurrency;
 using BenchmarkDotNet.Attributes;
 
-namespace App.Metrics.Benchmarks.ConcurrencyBenchmarks
+namespace App.Metrics.Benchmarks.BenchmarkDotNetBenchmarks.Concurrency
 {
-    public class PaddedAtomicLongBenchmark : DefaultBenchmarkBase
+    public class AtomicLongArrayBenchmark : DefaultBenchmarkBase
     {
-        private PaddedAtomicLong _num;
+        private AtomicLongArray _num;
 
         [Setup]
         public override void Setup()
         {
-            _num = new PaddedAtomicLong();
+            _num = new AtomicLongArray(10);
         }
 
         [Benchmark]
@@ -25,7 +25,7 @@ namespace App.Metrics.Benchmarks.ConcurrencyBenchmarks
         [Benchmark]
         public void Get()
         {
-            var x = _num.GetValue();
+            var x = _num.GetValue(1);
         }
 
         [Benchmark]
