+++ benchmarks/App.Metrics.Benchmarks/BenchmarkDotNetBenchmarks/Concurrency/AtomicLongBenchMark.cs
@@ -1,25 +1,25 @@
-// Copyright (c) Allan Hardy. All rights reserved.
+﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using App.Metrics.Concurrency;
 using BenchmarkDotNet.Attributes;
 
-namespace App.Metrics.Benchmarks.ConcurrencyBenchmarks
+namespace App.Metrics.Benchmarks.BenchmarkDotNetBenchmarks.Concurrency
 {
-    public class StripedLongAdderBenchmark : DefaultBenchmarkBase
+    public class AtomicLongBenchmark : DefaultBenchmarkBase
     {
-        private StripedLongAdder _num;
+        private AtomicLong _num;
 
         [Setup]
         public override void Setup()
         {
-            _num = new StripedLongAdder();
+            _num = new AtomicLong();
         }
 
         [Benchmark]
         public void Decrement()
         {
-            _num.Decrement(1);
+            _num.Decrement();
         }
 
         [Benchmark]
@@ -31,7 +31,7 @@ namespace App.Metrics.Benchmarks.ConcurrencyBenchmarks
         [Benchmark]
         public void Increment()
         {
-            _num.Increment(1);
+            _num.Increment();
         }
     }
 }
