+++ benchmarks/App.Metrics.Benchmarks/BenchmarkDotNetBenchmarks/Concurrency/ThreadLocalLongAdderBenchMark.cs
@@ -1,19 +1,19 @@
-﻿// Copyright (c) Allan Hardy. All rights reserved.
+// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using App.Metrics.Concurrency;
 using BenchmarkDotNet.Attributes;
 
-namespace App.Metrics.Benchmarks.ConcurrencyBenchmarks
+namespace App.Metrics.Benchmarks.BenchmarkDotNetBenchmarks.Concurrency
 {
-    public class AtomicLongArrayBenchmark : DefaultBenchmarkBase
+    public class ThreadLocalLongAdderBenchmark : DefaultBenchmarkBase
     {
-        private AtomicLongArray _num;
+        private ThreadLocalLongAdder _num;
 
         [Setup]
         public override void Setup()
         {
-            _num = new AtomicLongArray(10);
+            _num = new ThreadLocalLongAdder();
         }
 
         [Benchmark]
@@ -25,7 +25,7 @@ namespace App.Metrics.Benchmarks.ConcurrencyBenchmarks
         [Benchmark]
         public void Get()
         {
-            var x = _num.GetValue(1);
+            var x = _num.GetValue();
         }
 
         [Benchmark]
