+++ benchmarks/App.Metrics.Benchmarks/XunitHarness/Concurrency.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-using App.Metrics.Benchmarks.ConcurrencyBenchmarks;
+using App.Metrics.Benchmarks.BenchmarkDotNetBenchmarks.Concurrency;
 using BenchmarkDotNet.Running;
 using Xunit;
