+++ benchmarks/App.Metrics.Benchmarks/XunitHarness/Counter.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
+using App.Metrics.Benchmarks.BenchmarkDotNetBenchmarks.Metrics;
 using App.Metrics.Benchmarks.Support;
 using Xunit;
 using Xunit.Abstractions;
