+++ benchmarks/App.Metrics.Benchmarks/XunitHarness/ReservoirSampling.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
+using App.Metrics.Benchmarks.BenchmarkDotNetBenchmarks.Sampling;
 using App.Metrics.Benchmarks.Support;
 using Xunit;
 using Xunit.Abstractions;
